(feat) Provide function for obtaining default config tree in test code
