+++ packages/framework/esm-utils/src/index.ts
@@ -2,6 +2,7 @@ export * from './age-helpers';
 export * from './omrs-dates';
 export * from './shallowEqual';
 export * from './storage';
+export * from './test-helpers';
 export * from './translate';
 export * from './version';
 export * from './retry';
